refactor(docs): checkbox-group dx
