+++ apps/docs/content/components/checkbox-group/disabled.raw.jsx
@@ -1,0 +1,13 @@
+import {CheckboxGroup, Checkbox} from "@nextui-org/react";
+
+export default function App() {
+  return (
+    <CheckboxGroup isDisabled defaultValue={["buenos-aires", "london"]} label="Select cities">
+      <Checkbox value="buenos-aires">Buenos Aires</Checkbox>
+      <Checkbox value="sydney">Sydney</Checkbox>
+      <Checkbox value="san-francisco">San Francisco</Checkbox>
+      <Checkbox value="london">London</Checkbox>
+      <Checkbox value="tokyo">Tokyo</Checkbox>
+    </CheckboxGroup>
+  );
+}
